Langchain Projects/requirements.txt
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.29

python-dotenv==1.0.1
openai==1.58.1
